doc: update docs
