+++ packages/shared/config/theme.ts
@@ -1,4 +1,4 @@
-import { NetworkThemeConfig } from '../model';
+import { Network, NetworkThemeConfig } from '../model';
 import dark from '../theme/antd/dark.json';
 import light from '../theme/antd/light.json';
 import vars from '../theme/antd/vars.json';
@@ -67,3 +67,26 @@ export enum THEME {
   LIGHT = 'light',
   DARK = 'dark',
 }
+
+export const chainColors: { [key in Network]: string } = {
+  'BNB Chain': '#f1b90e',
+  'crab-dvm': '#B32BB6',
+  'crab-parachain': '#512dbc',
+  'darwinia-dvm': '#FF007A',
+  'pangolin-dvm': '#3083dd',
+  'pangolin-parachain': '#4b30dd',
+  'pangoro-dvm': '#07488f',
+  arbitrum: '#2c384b',
+  astar: '#1b8ff8',
+  avalanche: '#e74140',
+  crab: '#cd201f',
+  darwinia: '#FF007A',
+  ethereum: '#1C87ED',
+  heco: '#05933a',
+  optimism: '#fe0421',
+  pangolin: '#4b30dd',
+  pangoro: '#07488f',
+  polkadot: '#f19135',
+  polygon: '#8447f3',
+  ropsten: '#c6c6c6',
+};
